This applies changes to allow witnessing transactions that match cardano-cli.
This change allows witnesses to be marshalled to CBOR in a manner that
matches cardano-cli. As well a fix is made to how transaction bodies
are marshalled to CBOR to make them match how cardano-cli does this so
that we can successfully witness a transaction and end up with the
same output. This required adding a tag to the transaction inputs
array.